requirements.txt
openai==2.24.0
python-dotenv==1.2.1
pdfplumber==0.11.9
pandas==3.0.1
tqdm==4.67.3
requests==2.31.0